requirements.txt
grpcio==1.73.1
grpcio-reflection==1.73.1
grpcio-tools==1.73.1
protobuf==6.31.1
